Text, button, titles added to basic layout
Remove the image inside left-card div and replaced with background-image
property.
Remove the Css changing the display to none and viceversa to allow the
image change with the media queries. Now only the background-image property
is being used
Change the width of the main card div
Add Text, button, and title to the card-right div. A flexbox might be needed
for centering purposes